roptuna: Hyperparameter Optimization Using the Optuna Framework

An R implementation of the Optuna hyperparameter optimization framework. Provides define-by-run search spaces, TPE sampling, trial pruning, and persistent study storage via SQLite.

See also

Author

Maintainer: Kailas Venkitasubramanian kailasv@gmail.com